+++ src/entities/snake_segment.rs
@@ -21,7 +21,7 @@ impl SnakeSegment {
                     ..default()
                 },
                 transform: Transform::from_xyz(x, y, -1.)
-                    .with_scale(Vec3::splat(PIXEL_SIZE as f32)),
+                    .with_scale(Vec3::splat(PIXEL_SIZE)),
                 ..default()
             },
             b: Segment { lifetime },
